+++ src/re_com/config.cljs
@@ -15,6 +15,10 @@
 
 (goog-define root-url-for-compiler-output "")
 
+(goog-define debug-parts? true)
+
+(goog-define log-format "js")
+
 ;; When include-args-desc? is true, the data structures for arguments and parts will be included in the output JS file,
 ;; otherwise they will not be included.
 (def ^boolean include-args-desc?